requirements.txt
absl-py==1.4.0
altair==4.1.0
astunparse==1.6.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.0.4
entrypoints==0.4
exceptiongroup==1.1.3
flatbuffers==23.5.26
gast==0.5.4
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.57.0
h11==0.14.0
h5py==3.9.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
numpy==1.22.4
oauthlib==3.2.2
opencv-python==4.5.5.64
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.1
pandas==1.4.4
Pillow==8.4.0
protobuf==3.19.6
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pympler==1.0.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.9.2
rsa==4.9
scikit-learn==1.1.3
scipy==1.11.2
selenium==4.11.2
semver==3.0.1
six==1.16.0
sklearn==0.0.post7
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
streamlit==1.9.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.4
tensorflow-estimator==2.8.0
tensorflow-io-gcs-filesystem==0.33.0
termcolor==2.3.0
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.64.1
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
validators==0.21.2
watchdog==3.0.0
Werkzeug==2.3.7
wrapt==1.15.0
wsproto==1.2.0
zipp==3.16.2
